prep: Explode data.js into 3 separate files
In order to make them easier to read
