+++ src/utils/data/index.js
@@ -1,0 +1,3 @@
+export * from "./pdf";
+export * from "./json";
+export * from "./studyData";
